% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLocationModel.R
\name{runLocationModel}
\alias{runLocationModel}
\title{Fits GAM model}
\usage{
runLocationModel(hospital_name, input_df)
}
\arguments{
\item{hospital_name}{Character with hospital name that is consistent with the hospitals available in input_df column "hospital"}

\item{input_df}{Input data.frame for model. Data.frame needs columns hospital (`character`), date (`Date`), hour (`integer`)}
}
\description{
Fits GAM model
}
