% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffingModels-package.R
\docType{package}
\name{staffingModels-package}
\alias{staffingModels}
\alias{staffingModels-package}
\title{staffingModels: Predictive staffing using generalized additive models (GAM)}
\description{
Patient volume based predictive staffing.
}
\author{
\strong{Maintainer}: Abbas Rizvi \email{arizvi@cbuscollaboratory.com}

}
\keyword{internal}
